requirements.txt
streamlit
pdfplumber
python-docx
spacy
nltk
sentence-transformers
openai